Moved finder into finder package
